move trade button
